# openapi.yaml
...
openapi: 3.0.3
info:
  title: Panuval Bookstore API
  version: 1.0.0
  description: API specification for the Panuval Tamil Bookstore mobile application.
servers:
  - url: https://api.panuval.com/v1
    description: Production Server
  - url: https://staging-api.panuval.com/v1
    description: Staging Server
paths:
...
  /app/home/layout:
    get:
      summary: Get Home Page Layout Structure
      description: Retrieves the metadata (type, order, visibility, ID) for sections to be displayed on the home page, ordered correctly. Content is loaded separately.
      responses:
        '200':
          description: Successfully retrieved home page layout structure.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectionMetadata'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /app/home/section/{sectionId}:
    get:
      summary: Get Home Page Section Content
      description: Retrieves the detailed content for a specific home page section, identified by its ID. Used for lazy loading.
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The unique ID of the section whose content is requested.
      responses:
        '200':
          description: Successfully retrieved section content. The structure varies based on the section's type.
          content:
            application/json:
              schema:
                # Content varies; use oneOf or examples to illustrate possibilities
                # Example showing possible responses based on sectionType
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ItemSummary'
                  - type: array
                    items:
                      $ref: '#/components/schemas/AuthorPublisherCategoryInfo'
                  - type: array
                    items:
                      $ref: '#/components/schemas/BannerContent'
                  - type: array
                    items:
                      $ref: '#/components/schemas/LabelContent'
        '404':
          $ref: '#/components/responses/NotFound' # If sectionId is invalid
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        errorCode:
          type: string
          description: An optional machine-readable error code.
          nullable: true
      required:
        - message
...
    AuthorPublisherCategoryInfo:
      type: object
      properties:
        name:
          type: string
          description: Name of the author, publisher, or category.
        imageUrl:
          type: string
          format: url
          nullable: true
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
      required:
        - name
    BannerContent:
      type: object
      properties:
        name:
          type: string
          description: Name of to display for the banner.
        imageUrl:
          type: string
          format: url
        linkType:
          type: string
          enum: [ITEM, NONE, SEARCH_FILTER]
        itemId:
          type: integer
          nullable: true
          description: ID of the item to link to, if applicable.  
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
          nullable: true
      required:
        - imageUrl
        - linkType
    LabelContent:
      type: object
      properties:
        name:
          type: string
          description: Name of to display for the Label.
        labelColor:
          type: string
          description: Color of the label.
        textColor:
          type: string
          description: Color of the text.
        linkType:
          type: string
          enum: [ITEM, NONE, SEARCH_FILTER]
        itemId:
          type: integer
          nullable: true
          description: ID of the item to link to, if applicable.  
        searchFilter:
          $ref: '#/components/schemas/SearchFilter'
          nullable: true
      required:
        - name
        - linkType
    SectionMetadata:
      type: object
      properties:
        sectionId:
          type: string
          format: uuid
        displayType:
          type: string
          enum: [TEXT, CAROUSEL, TILE]
        contentType: # This is the type of content in the section
          type: string
          enum: [LABEL, BANNER, BOOK, CATEGORY, AUTHOR, PUBLISHER]
        title:
          type: string
          nullable: true
        showTitle:
          type: boolean
          description: Indicates if the title should be displayed.
          default: true
        subTitle:
          type: string
          nullable: true
        showSubTitle:
          type: boolean
          description: Indicates if the subtitle should be displayed.
          default: true
        showViewAll:
          type: boolean
          description: Indicates if "View All" should be displayed.
          default: false
        showName:
          type: boolean
          description: Indicates if the name of the items to  be displayed.
          default: true
        showAuthor:
          type: boolean
          description: Indicates if the author of the items to be displayed, if present.
          default: true
        order:
          type: integer
        visible:
          type: boolean
      required:
        - sectionId
        - displayType
        - order
        - visible
    # Need ItemSummary for book sections later
...
    ItemSummary:
      type: object
      properties:
        itemId:
          type: integer
        name:
          type: string
          description: Name of the item.
        imageUrl:
          type: string
          format: url
          nullable: true
        price:
          type: number
          format: float
          description: Current selling price.
        originalPrice:
          type: number
          format: float
          description: Original price, MRP.
        discountPercentage:
          type: number
          format: float
          description: Percentage discount applied. Null if no discount.
          nullable: true
        stockStatus:
          $ref: '#/components/schemas/StockStatus'
        label:
          $ref: '#/components/schemas/ItemLabel'
      required:
        - itemId
        - title
        - price
        - stockStatus
...
    StockStatus:
      type: string
      enum: [AVAILABLE, IN_STOCK, OUT_OF_STOCK, PRE_ORDER, OUT_OF_PRINT]
      description: Current stock status of the item
...
    ItemLabel:
      type: object
      properties:
        label:
          type: string
          description: Label for the item, such as "Best Seller", "New Arrival", etc.
        showLabel:
          type: boolean
          description: Indicates if the label should be displayed.
          default: true 
        labelType:
          type: string
          enum: [BEST_SELLER, NEW_ARRIVAL, AWARD_WINNING, NONE]
          description: Type of label to be displayed.
    SearchFilter:
      type: string
      description: A string representing the search filter parameters.
      example: 'authorId=123&categoryId=456'
...
  responses:
...
    NotFound:
      description: Not Found - The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth: # Can be named anything, used later to reference
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, for documentation
      description: "JWT token obtained after successful OTP verification."